Move contributors list component to its own file to improve readability.
