+++ components/ContributorsList.tsx
@@ -1,0 +1,71 @@
+import styled from '@emotion/styled';
+import * as React from 'react';
+
+import TopBar from './shared/TopBar';
+
+const ContributorsListWrapper = styled.div`
+  background-color: #feecde;
+  height: 461px;
+  max-width: 408px;
+  overflow: scroll;
+`;
+
+const ContributorsListTitle = styled.h2`
+  margin: 20px;
+  font-family: Terminal;
+  font-size: 30px;
+`;
+
+const AliasesWrapper = styled.div`
+  column-count: 2;
+  margin: 0px 20px;
+`;
+
+const Alias = styled.div`
+  margin-bottom: 10px;
+  font-size: 14px;
+  font-family: Favorit Pro;
+`;
+
+const ContributorsList = (): JSX.Element => {
+  const aliases = [
+    '@mindapi',
+    '@fakepixels',
+    '@zaz',
+    '@scottsgc',
+    '@fullyallocated',
+    '@tsully',
+    '@mindapi',
+    '@fakepixels',
+    '@zaz',
+    '@scottsgc',
+    '@fullyallocated',
+    '@tsully',
+    '@mindapi',
+    '@fakepixels',
+    '@zaz',
+    '@scottsgc',
+    '@fullyallocated',
+    '@tsully',
+    '@mindapi',
+    '@fakepixels',
+    '@zaz',
+    '@scottsgc',
+    '@fullyallocated',
+    '@tsully',
+  ];
+
+  return (
+    <ContributorsListWrapper>
+      <TopBar mandatory={false} />
+      <ContributorsListTitle>CONTRIBUTOR DIRECTORY</ContributorsListTitle>
+      <AliasesWrapper>
+        {aliases.map((alias, i) => (
+          <Alias key={'alias' + i}>{alias}</Alias>
+        ))}
+      </AliasesWrapper>
+    </ContributorsListWrapper>
+  );
+};
+
+export default ContributorsList;
