+++ pages/[orgName]/index.tsx
@@ -5,13 +5,13 @@ import { ethers } from 'ethers';
 import { useRouter } from 'next/router';
 import { useWeb3, useContracts } from '../../shared/contexts';
 
+import ContributorsList from '../../components/ContributorsList';
 import OrgSummary from '../../components/OrgSummary';
 import Button from '../../components/shared/Button';
 import ContributionModal from '../../components/Contributions/ContributionsModal';
 import ClaimRewardsModal from '../../components/ClaimRewards/ClaimRewardsModal';
 import GiveRewardsModal from '../../components/GiveRewards/GiveRewardsModal';
 import Layout from '../../components/shared/Layout';
-import TopBar from '../../components/shared/TopBar';
 
 import defaultOSFactoryABI from '../../utils/abi/DefaultOSFactory.json';
 import defaultOSABI from '../../utils/abi/DefaultOS.json';
@@ -173,71 +173,6 @@ const ContentWrapper = styled.div`
     padding: 10px;
   }
 `;
-
-const ContributorsListWrapper = styled.div`
-  background-color: #feecde;
-  height: 461px;
-  max-width: 408px;
-  overflow: scroll;
-`;
-
-const ContributorsListTitle = styled.h2`
-  margin: 20px;
-  font-family: Terminal;
-  font-size: 30px;
-`;
-
-const AliasesWrapper = styled.div`
-  column-count: 2;
-  margin: 0px 20px;
-`;
-
-const Alias = styled.div`
-  margin-bottom: 10px;
-  font-size: 14px;
-  font-family: Favorit Pro;
-`;
-
-const ContributorsList = (): JSX.Element => {
-  const aliases = [
-    '@mindapi',
-    '@fakepixels',
-    '@zaz',
-    '@scottsgc',
-    '@fullyallocated',
-    '@tsully',
-    '@mindapi',
-    '@fakepixels',
-    '@zaz',
-    '@scottsgc',
-    '@fullyallocated',
-    '@tsully',
-    '@mindapi',
-    '@fakepixels',
-    '@zaz',
-    '@scottsgc',
-    '@fullyallocated',
-    '@tsully',
-    '@mindapi',
-    '@fakepixels',
-    '@zaz',
-    '@scottsgc',
-    '@fullyallocated',
-    '@tsully',
-  ];
-
-  return (
-    <ContributorsListWrapper>
-      <TopBar mandatory={false} />
-      <ContributorsListTitle>CONTRIBUTOR DIRECTORY</ContributorsListTitle>
-      <AliasesWrapper>
-        {aliases.map((alias, i) => (
-          <Alias key={'alias' + i}>{alias}</Alias>
-        ))}
-      </AliasesWrapper>
-    </ContributorsListWrapper>
-  );
-};
 
 const Home = (): JSX.Element => {
   const router = useRouter();
